requirements.txt
iniherit==0.3.3
Momoko==2.2.0
psycopg2==2.6.1
pudb==2015.3
Pygments==2.0.2
python-slugify==1.1.4
redis==2.10.3
six==1.9.0
sockjs-tornado==1.0.2
tornado==4.2.1
tornado-redis==2.4.18
Unidecode==0.4.18
urwid==1.3.0
wheel==0.24.0
yoyo-migrations==5.0.0.dev0
